Purpose
The purpose of this project is to get all details of the user's RC-Book(Registration Certificate) for bike/car and generate the QR-Code from it.
Once QR is generated then generate file is stored in .png format in your same folder. Then you can print the same file and put it on your bike/car in a safe & easy to access place.